Implement LAHDMod to force menu scaling.
This implements "PageManager.lahdmod" which allows overriding the scaling used for the various menus in the game including the file selection screen and options menu screen. Like other menus, it allows entering fractional values instead of integers, which will distort the graphics but may be preferrable on small screens.

+++ ladxhd_game_source_code/ProjectZ.Core/InGame/Interface/InterfacePage.cs
@@ -23,7 +23,7 @@ namespace ProjectZ.InGame.Interface
             PageLayout?.Update();
         }
 
-        public virtual void Draw(SpriteBatch spriteBatch, Vector2 position, int scale, float transparency)
+        public virtual void Draw(SpriteBatch spriteBatch, Vector2 position, float scale, float transparency)
         {
             PageLayout?.Draw(spriteBatch, position, scale, transparency);
         }
