+++ ladxhd_game_source_code/ProjectZ.Core/InGame/Pages/Settings/ReduxSettingsPage.cs
@@ -193,7 +193,7 @@ namespace ProjectZ.InGame.Pages
             return " " + Game1.LanguageManager.GetString(option, "error");
         }
 
-        public override void Draw(SpriteBatch spriteBatch, Vector2 position, int height, float alpha)
+        public override void Draw(SpriteBatch spriteBatch, Vector2 position, float height, float alpha)
         {
             // Always draw the menu even when not showing tooltips.
             base.Draw(spriteBatch, position, height, alpha);
